requirements.txt
fastapi
uvicorn[standard]
pandas
requests
python-dotenv
xgboost
scikit-learn
yfinance